// client/src/components/signup/signup_interfaces.ts
export enum InvalidSignupErrors {
    FIRST_NAME = 'First name is invalid.',
    LAST_NAME = 'Last name is invalid.',
    EMAIL = 'Email is invalid.',
    PASSWORD = 'Password is invalid.',
    PHONE = 'Phone number is invalid.',
    BIRTH_DATE = 'Date of birth is invalid.',
    STREET_ADDRESS = 'Street address is invalid.',
    POST_CODE = 'Post code is invalid.',
    CITY = 'City is invalid',
    COUNTRY = 'Country is invalid.',
}